renaming xpbarrel namespace
